Moved components into views
