+++ client/src/views/admin-add-course.vue
@@ -62,7 +62,7 @@
 
 
 <script>
-import Header from '../components/admin-header.vue'
+import Header from './components/admin-header.vue'
 
 export default {
   components: {
