+++ client/src/views/admin-manage-courses.vue
@@ -38,7 +38,7 @@
 </template>
 
 <script>
-import Header from '../components/admin-header.vue'
+import Header from './components/admin-header.vue'
 
 export default {
   components: {
