+++ client/src/views/student-login.vue
@@ -1,30 +1,33 @@
-<template >
+<template>
 <div class="main">
-        <div class="login-page">
+    <div class="login-page">
         <div class="form">
-        <form class="login-form">
-            <input id="idnum" type="number" placeholder="ID Number"/>
-            <input id="password" type="password" placeholder="Password"/>
-            <button type="submit" id="login"><b>login</b></button>
-            <p class="message">Not registered? <router-link :to="{name: 'Register'}">Create an account</router-link></p>
-            <!-- <p class="error"></p> -->
-        </form>
+            <form class="login-form" @submit.prevent>
+                <input id="idnum" type="number" placeholder="ID Number" />
+                <input id="password" type="password" placeholder="Password"/>
+                <router-link :to="{name: 'Search'}"><button type="submit" id="login"><b>login</b></button></router-link>
+                <p class="message">Not registered? <router-link :to="{name: 'Register'}">Create an account</router-link></p>
+            </form>
         </div>
     </div>
     <div class="footer"> 
-   
-        <p class="message"> <router-link :to="{name: 'loginAdmin'}">Login as Admin </router-link> </p>
-
-       
-         
+        <p class="message"> <router-link :to="{name: 'loginAdmin'}">Login as Admin </router-link> </p>     
     </div>
 </div>
 </template>
 
 
 <script>
-
 export default {
+    name: 'Login',
+    data() {
+        return {
+            input: {
+                username: '',
+                password: ''
+            }
+        }
+    }
 }
 </script>
