+++ client/src/views/student-register.vue
@@ -2,8 +2,8 @@
 <div class="main">
   <div class="register-page">
     <div class="form">
-      <form class="register-form">
-        <input id="fullname" type="text" placeholder="Full Name (First Middle Last)"/>
+      <form class="register-form" @submit.prevent>
+        <input id="fullname" type="text" placeholder="Full Name"/>
         <input id="idnum" type="number" placeholder="ID Number"/>
 
         <input id="email" type="email" placeholder="E-mail Address"/>
@@ -25,7 +25,8 @@
 
 <script>
 export default {
-  
+    name: "RegisterForm",
+
 }
 </script>
